Reverted the tfmaps to storing floats instead of unsigned shorts, because the mapping to ushorts sucks if you don't know your potential max and min values. Methinks the whole tfmap system has to be rethought in accordance with the DIIC3 gui: ending up with a pixmap is clearly not ideal. Vectorized objects, such as the Matlab-style patches, would be far more flexible.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@193 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/libmptk/mp_types.h
@@ -55,7 +55,8 @@ typedef double MP_Real_t;   /* Inner products data type */
 typedef char   MP_Bool_t;         /* Boolean data type */
 typedef char   MP_Bool_On_Disk_t; /* Disk version of Boolean data type */
 
-typedef unsigned short int MP_Tfmap_t;  /* Type of the Tfmap data */
+//typedef unsigned short int MP_Tfmap_t;  /* Type of the Tfmap data */
+typedef float  MP_Tfmap_t;  /* Type of the Tfmap data */
 #define TFMAP_NUM_DISCRETE_LEVELS 65536 /* Number of levels when discretizing
 					   real values in a tfmap */
